# backend/contracts/analytics-page.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Free Page Documents",
    "version": "free-page/v1",
    "description": "Independent HTML page assets. Not BoardSpec. D6=PATCH confirm, D9=SAVE."
  },
...
  "paths": {
    "/api/v1/analytics/page-documents/previews": {
      "post": {
        "operationId": "page_generate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Page preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/previews/{preview_id}": {
      "get": {
        "operationId": "page_preview",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/previews/{preview_id}/confirm": {
      "post": {
        "operationId": "page_confirm",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committed snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/previews/{preview_id}/cancel": {
      "post": {
        "operationId": "page_cancel",
        "parameters": [
          {
            "name": "preview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCancelResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/page-documents/pages/{page_id}": {
      "get": {
        "operationId": "page_get",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committed snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/pages/{page_id}/versions": {
      "get": {
        "operationId": "page_history",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PageRevision"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/pages/{page_id}/patch-preview": {
      "post": {
        "operationId": "page_patch",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagePatchPreview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Page preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/pages/{page_id}/save-preview": {
      "post": {
        "operationId": "page_save",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageSavePreview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Page preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/page-documents/pages/{page_id}/rollback-preview": {
      "post": {
        "operationId": "page_rollback",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageRollbackPreview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Page preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePreview"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PageResource": {
        "additionalProperties": false,
        "properties": {
          "resource_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Resource Id",
            "type": "string"
          },
          "content_type": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._+-]+/[A-Za-z0-9._+-]+$",
            "title": "Content Type",
            "type": "string"
          },
          "sha256": {
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "byte_length": {
            "maximum": 2000000,
            "minimum": 0,
            "title": "Byte Length",
            "type": "integer"
          }
        },
        "required": [
          "resource_id",
          "content_type",
          "sha256",
          "byte_length"
        ],
        "title": "PageResource",
        "type": "object"
      },
      "PageNodeMapEntry": {
        "additionalProperties": false,
        "properties": {
          "node_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Node Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "static_element",
              "dynamic_region",
              "whole_page"
            ],
            "title": "Kind",
            "type": "string"
          },
          "selector": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Selector",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "kind",
          "selector"
        ],
        "title": "PageNodeMapEntry",
        "type": "object"
      },
      "PagePackage": {
        "additionalProperties": false,
        "properties": {
          "html": {
            "maxLength": 1048576,
            "minLength": 1,
            "title": "Html",
            "type": "string"
          },
          "css": {
            "default": "",
            "maxLength": 524288,
            "title": "Css",
            "type": "string"
          },
          "js": {
            "default": "",
            "maxLength": 524288,
            "title": "Js",
            "type": "string"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/PageResource"
            },
            "maxItems": 32,
            "title": "Resources",
            "type": "array"
          },
          "node_map": {
            "items": {
              "$ref": "#/components/schemas/PageNodeMapEntry"
            },
            "maxItems": 2000,
            "title": "Node Map",
            "type": "array"
          }
        },
        "required": [
          "html"
        ],
        "title": "PagePackage",
        "type": "object"
      },
      "PageBinding": {
        "additionalProperties": false,
        "properties": {
          "binding_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Binding Id",
            "type": "string"
          },
          "result_ref": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Result Ref",
            "type": "string"
          },
          "node_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Node Id"
          },
          "data_ref": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Data Ref"
          },
          "mode": {
            "default": "summary",
            "enum": [
              "summary",
              "page",
              "range"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "binding_id",
          "result_ref"
        ],
        "title": "PageBinding",
        "type": "object"
      },
      "PageBindingManifest": {
        "additionalProperties": false,
        "properties": {
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/PageBinding"
            },
            "maxItems": 64,
            "title": "Bindings",
            "type": "array"
          },
          "result_refs": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
              "type": "string"
            },
            "maxItems": 64,
            "title": "Result Refs",
            "type": "array"
          }
        },
        "title": "PageBindingManifest",
        "type": "object"
      },
      "PageDraft": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Title",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "package": {
            "$ref": "#/components/schemas/PagePackage"
          },
          "binding_manifest": {
            "$ref": "#/components/schemas/PageBindingManifest"
          },
          "origin_path": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string",
            "title": "Origin Path"
          },
          "origin_file_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "type": "string",
            "title": "Origin File Id"
          }
        },
        "required": [
          "title",
          "session_id",
          "package"
        ],
        "title": "PageDraft",
        "type": "object"
      },
      "PageDocument": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Title",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "package": {
            "$ref": "#/components/schemas/PagePackage"
          },
          "binding_manifest": {
            "$ref": "#/components/schemas/PageBindingManifest"
          },
          "origin_path": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string",
            "title": "Origin Path"
          },
          "origin_file_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "type": "string",
            "title": "Origin File Id"
          },
          "schema_version": {
            "const": "free-page/v1",
            "default": "free-page/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "page_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Page Id",
            "type": "string"
          },
          "version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Version",
            "type": "integer"
          },
          "binding_state": {
            "enum": [
              "UNBOUND_SAMPLE",
              "BOUND_VERIFIED",
              "BOUND_STALE"
            ],
            "title": "Binding State",
            "type": "string"
          }
        },
        "required": [
          "title",
          "session_id",
          "package",
          "page_id",
          "version",
          "binding_state"
        ],
        "title": "PageDocument",
        "type": "object"
      },
      "PagePatchPreview": {
        "additionalProperties": false,
        "description": "D6: patch preview of source and/or manifest. Confirm is a separate call.",
        "properties": {
          "base_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Base Version",
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string",
            "title": "Title"
          },
          "package": {
            "$ref": "#/components/schemas/PagePackage"
          },
          "binding_manifest": {
            "$ref": "#/components/schemas/PageBindingManifest"
          }
        },
        "required": [
          "base_version"
        ],
        "title": "PagePatchPreview",
        "type": "object"
      },
      "PageSavePreview": {
        "additionalProperties": false,
        "description": "D9: explicit save of the host in-memory draft. Exit-edit is not save.",
        "properties": {
          "base_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Base Version",
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Title",
            "type": "string"
          },
          "package": {
            "$ref": "#/components/schemas/PagePackage"
          },
          "binding_manifest": {
            "$ref": "#/components/schemas/PageBindingManifest"
          }
        },
        "required": [
          "base_version",
          "title",
          "package",
          "binding_manifest"
        ],
        "title": "PageSavePreview",
        "type": "object"
      },
      "PageRollbackPreview": {
        "additionalProperties": false,
        "properties": {
          "base_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Base Version",
            "type": "integer"
          },
          "to_version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "To Version",
            "type": "integer"
          }
        },
        "required": [
          "base_version",
          "to_version"
        ],
        "title": "PageRollbackPreview",
        "type": "object"
      },
      "PageSnapshot": {
        "additionalProperties": false,
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/PageDocument"
          }
        },
        "required": [
          "spec"
        ],
        "title": "PageSnapshot",
        "type": "object"
      },
      "PagePreview": {
        "additionalProperties": false,
        "properties": {
          "preview_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Preview Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "APPLIED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          },
          "operation": {
            "enum": [
              "GENERATE",
              "PATCH",
              "SAVE",
              "ROLLBACK"
            ],
            "title": "Operation",
            "type": "string"
          },
          "base_version": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "title": "Base Version",
            "type": "integer"
          },
          "expires_at_ms": {
            "title": "Expires At Ms",
            "type": "integer"
          },
          "snapshot": {
            "$ref": "#/components/schemas/PageSnapshot"
          }
        },
        "required": [
          "preview_id",
          "status",
          "operation",
          "base_version",
          "expires_at_ms",
          "snapshot"
        ],
        "title": "PagePreview",
        "type": "object"
      },
      "PageRevision": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Version",
            "type": "integer"
          },
          "operation": {
            "enum": [
              "GENERATE",
              "PATCH",
              "SAVE",
              "ROLLBACK"
            ],
            "title": "Operation",
            "type": "string"
          },
          "created_at_ms": {
            "title": "Created At Ms",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "operation",
          "created_at_ms"
        ],
        "title": "PageRevision",
        "type": "object"
      },
...
      "PageCancelResult": {
        "additionalProperties": false,
        "properties": {
          "preview_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]{1,128}$",
            "title": "Preview Id",
            "type": "string"
          },
          "status": {
            "const": "CANCELLED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "preview_id",
          "status"
        ],
        "title": "PageCancelResult",
        "type": "object"
      },
...
    }
  },
...
}